requirements.txt
unicorn-binance-websocket-api==1.28.0
python-telegram-bot==13.1
python-dotenv==0.15.0
pandas==1.1.3
python-binance==0.7.5
pyyaml==5.3.1